refactor: Delete presenter and add functions in controller
